+++ RobotCodeRestart/src/org/usfirst/frc/team4627/robot/subsystems/Sensors.java
@@ -11,8 +11,8 @@ import edu.wpi.first.wpilibj.command.Subsystem;
 public class Sensors extends Subsystem {
 	
 	ADXRS450_Gyro gyroSPI = new ADXRS450_Gyro();
-	public double preAngle;
-	public double postAngle;
+	public int preAngle;
+	public int postAngle;
 	
 	
 	public double getGyroAngle() {
